requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
python-dotenv==1.2.1
requests==2.32.5
gspread==6.2.1
google-auth==2.48.0
Pillow==12.1.1
pytesseract==0.3.13